fix access token authentication

+++ Model/Resolver/Delete/AbstractResolver.php
@@ -40,6 +40,10 @@ class AbstractResolver extends \Mageplaza\GiftWrapGraphQl\Model\Resolver\Abstrac
      */
     protected function handleArgs(array $args)
     {
+        if (!$this->auth->isAllowed($args['accessToken'], $this->_aclResource)) {
+            throw new GraphQlInputException(__("The consumer isn't authorized to access %1", $this->_aclResource));
+        }
+
         try {
             return $this->filter->deleteEntity($args['id'], $this->_type);
         } catch (NoSuchEntityException $e) {
